Move Board component into components folder
